Removed bcrypt from Libraries section
Because of the large amount of dependencies required by bcrypt, we decided to replace it with Node's included crypto package. Specifically, we were running into many issues trying to build QueueUp on Windows machines because of the bcrypt dependencies. The crypto package offers us a way to encrypt data more specifically than bcrypt allowed.